Added detailed messages to settings cloning
Added ability to eject unpackager trackers
Added ability to clear recipe holders and markers in crafting grid
Changed saving settings to cloner to use shift-right click
Fixed Packager recipe holder copy and pasting

+++ src/main/java/thelm/packagedauto/client/gui/GuiItemAmountSpecifying.java
@@ -50,7 +50,7 @@ public class GuiItemAmountSpecifying extends GuiAmountSpecifying<ContainerItemAm
 			int amount = MathHelper.clamp(Integer.parseInt(amountField.getText()), 0, maxAmount);
 			ItemStack newStack = stack.copy();
 			newStack.setCount(amount);
-			PacketHandler.INSTANCE.sendToServer(new PacketSetItemStack((short)containerSlot, newStack));
+			PacketHandler.INSTANCE.sendToServer(new PacketSetItemStack(containerSlot, newStack));
 			close();
 		}
 		catch(NumberFormatException e) {
